Prevent scanning in error recovery
Tree-sitter v0.20.8 enters an infinite loop if an empty token is
returned while scanning in error recovery mode. We can prevent
that by not scanning at all in error recovery mode. This is only
a partial fix though: instead the scanner should be debugged for
why it is returning empty tokens during error recovery.